Committing current changes

+++ FinanceApp/FinanceApp.Services.Interfaces/IAccountService.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+
 using FinanceApp.Model;
 
 namespace FinanceApp.Service.Interface
@@ -6,9 +7,9 @@ namespace FinanceApp.Service.Interface
     public interface IAccountService
     {
         IEnumerable<AccountModel> GetAccounts();
-        AccountModel GetAccount(int id);
+        AccountModel GetAccount(long id);
         void CreateAccount(CreateAccountModel account);
         void UpdateAccount(UpdateAccountModel account);
-        void DeleteAccount(int id);
+        void DeleteAccount(long id);
     }
 }
